# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 관리의달인 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [vendors]
      summary: List quotes
      parameters:
        - name: rfqId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [submitted, accepted, rejected]
      responses:
        "200":
          description: Quote list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [vendors]
      summary: Submit a quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [vendors]
      summary: Get a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
    patch:
      operationId: updateQuote
      tags: [vendors]
      summary: Update a quote status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found

  # ── Work Reports (작업완료보고) ──────────────────────────
...
components:
  schemas:
...
    Quote:
      type: object
      properties:
        id:
          type: integer
        rfqId:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        totalAmount:
          type: number
        itemBreakdown:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        availableDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [submitted, accepted, rejected]
        contractFilePath:
          type: ["string", "null"]
        contractUploadedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - rfqId
        - vendorId
        - vendorName
        - totalAmount
        - status
        - createdAt
        - updatedAt
    CreateQuoteBody:
      type: object
      properties:
        rfqId:
          type: integer
        vendorId:
          type: integer
        vendorName:
          type: string
        totalAmount:
          type: number
        itemBreakdown:
          type: ["string", "null"]
        scope:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        availableDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        requiredDocsComplete:
          type: boolean
          default: false
      required:
        - rfqId
        - vendorId
        - vendorName
        - totalAmount
    UpdateQuoteBody:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, accepted, rejected]
        notes:
          type: ["string", "null"]
        contractFilePath:
          type: ["string", "null"]
        contractUploadedAt:
          type: ["string", "null"]
          format: date-time
...