# lib/api-zod/src/generated/types/updateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 관리의달인 API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateQuoteBodyStatus } from "./updateQuoteBodyStatus";

export interface UpdateQuoteBody {
  status?: UpdateQuoteBodyStatus;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  contractFilePath?: string | null;
  /** @nullable */
  contractUploadedAt?: Date | null;
}